Struct c_vec::CSlice

pub struct CSlice<T> { /* fields omitted */ }

The type representing an 'unsafe' foreign chunk of memory

Methods

impl<T> CSlice<T>

pub unsafe fn new(base: *mut T, len: usize) -> CSlice<T>

Create a CSlice from a raw pointer to a buffer with a given length.

Panics if the given pointer is null. The returned vector will not attempt to deallocate the vector when dropped.

Arguments

base - A raw pointer to a buffer

len - The number of elements in the buffer

pub fn get<'a>(&'a self, ofs: usize) -> Option<&'a T>

Retrieves an element at a given index, returning None if the requested index is greater than the length of the vector.

pub fn get_mut<'a>(&'a mut self, ofs: usize) -> Option<&'a mut T>

Retrieves a mutable element at a given index, returning None if the requested index is greater than the length of the vector.

pub fn len(&self) -> usize

Returns the number of items in this vector.

pub fn is_empty(&self) -> bool

Returns whether this vector is empty.

Important traits for CSliceIter<'a, T>
impl<'a, T> Iterator for CSliceIter<'a, T> type Item = &'a T;

pub fn iter<'a>(&'a self) -> CSliceIter<'a, T>

Returns an iterator over CVec.